refactor(campaigns): extract TEXT_CAMPAIGN add to _campaigns_text.py (#602, step 6)
Closes #618

Part of #602

Extract the TEXT_CAMPAIGN add dispatch branch (~200 lines) from

campaigns.py into a dedicated _campaigns_text.py module with

build_add_block(). The update path stays in the shared text_family

block (TEXT + UNIFIED + DYNAMIC) and will be split in later steps.

CLI surface is byte-for-byte identical: all 502 offline tests pass

including dry-run golden tests, wsdl parity gate, and api coverage.
